Use jest in e2e tests
