--- e2e/fixtures/01-simple-readme/status.txt
@@ -1,1 +1,0 @@
-0
